Properly handle nested styles in layout construction
Previously, we used a stack-based algorithm to walk the parsed tree when
constructing the layout boxes. However, this made it tricky to keep
track of where we were within nested style blocks. This commit switches
to a recursive algorithm to properly handle nested styles and introduces
a few new structures to aid with keeping track of the construction state.