Fix spacing after mid-line font changes
Previously, we would emit the full style block as soon as it finished.
This meant that when a block didn't span the full line, as in a
mid-paragraph font change, the words before the change but on the same
line were assigned positions before the second half of the line.
Therefore, when assigning the spaces for justified text, only the second
half of the line would be altered.

Now, we wait until the entire line has been emitted to assign spaces.

Fixes: https://todo.sr.ht/~reesmichael1/burro/4