Allow document configuration section
We now have support for changing the margins, but that's just a way to
demonstrate the real meat of this commit: we can now parse commands
before the .start command to configure the document! We'll obviously
expand the commands allowed in the document configuration, but this is
good to get started.