Refactor page storage
Before, we were storing all pages in the pages vector and popping off
the last one as the "current" page whenever it was needed. This was
tricky to keep track of though, as it needed to be put back whenever it
was done, so there were blocks like

// pop page
// process(page)
// push page back <- don't forget this or we'll crash!

// process2 (this re-pops the page and hopefully puts it back)

My original design did use a current_page within the LayoutBuilder, but
it didn't work for #reasons. We're past those now though, so we've
changed to a saner design.