Add em and en dashes in tokenization stage
This required a slight update to the lexer: we no longer require hyphens
to be escaped in any context. The syntax "[-]" creates a reset token,
and anywhere else, the "-" produces a literal string.

It's generally a good idea to require as little escaping as possible
when the context makes it clear, so we'll aim to have more updates like
this one.