Initial support for hyphenation when justified
This is really a checkpoint commit rather than a finished product. The
next 3-4 commits that will focus on hyphenation. Those commits will
focus on improving:

* Currently, there is no way to limit the number of consecutive
  hyphenated words, which is a typographic sin. We'll add a parameter
  for maximum consecutive hyphenations.
* We'll also add a .spread command, which will break the current line
  while justifying its words. (In combination, these two features will
  allow for robust automatic hyphenation that can be easily overridden.)
* Justification still does not compress the spaces between words, only
  expands it.

This commit also wraps all of the TextUnits in Arcs, which is the
easiest way to create new words (such as a syllable followed by a
hyphen) on th fly. I have some ideas about how to restructure this, but
wanted to save a working state first.