Add .spread command
In combination with limited consecutive hyphenation, this gives a
typographer powerful control over what words are hyphenated in the
output.

This is enough work to close the hyphenation ticket. Next, we'll also
check if we should compress the spaces when hyphenating, which will let
us close #8.

Implements: https://todo.sr.ht/~reesmichael1/burro/7