Add command for altering letter spacing
I've also updated the demo to use Cardo instead of Bembo (since the free
version of Bembo I downloaded is missing ligatures).

Eventually, we'll add support for finding the smallcaps that are
supposedly in the Cardo OTF files, but for now, we stuck with using the
Bembo smallcap (which is a good chance to show off switching families in
the demo).