Begin working on horizontal rules
As of this commit, horizontal rules are "working," but they're still
very finicky. The bug with commands having to be in a certain order is
particularly annoying here. It makes handling the spacing around rules
very tricky.

This also finally adds an example of the curly brace syntax! This is how
we'll pass multiple options to commands.

I've thought about making the spacing around the rule a parameter via
the curly brace syntax, but decided against it for now. Mom doesn't
alter vertical spacing at all when inserting rules, which I think is a
good model to follow. This is open to be revisited once more documents
are set in Burro and we have case studies of the ergonomics.