Rework distinction between normal and style change commands
Now, any command can be entered at any point on the page, i.e., we no
longer forbid entering (most) commands mid-paragraph. This restriction
was becoming annoying in several ways. It made it impossible to do
things like changing the leading or margins in the middle of a
paragraph, and it meant that a document author had to be very careful
not to use commands in the wrong order, because a "style change" command
would start a new "paragraph," which would then reject a "normal"
command.

This also fixes the weirdness with adding spacing around rules. The
rules.bur example file has been cleaned up to match this.

It's possible that this commit introduces subtly wrong behavior, as we
don't have layout tests. All of the example files produce the correct
results, and we'll be on the lookout for issues in future documents.

Fixes: https://todo.sr.ht/~reesmichael1/burro/14