Fix error on comment at document beginning
This was caused by the lexer expecting all comments to be preceded by a
newline. Everywhere else in the document, this makes sense, because we
require comments to be at the beginning of the line. However, on the
very first line, there is obviously no newline to be followed!

There might be a better way to handle this, but for now, we just check
if there's a comment before we even start the lexing process and discard
it if present.

Fixes: https://todo.sr.ht/~reesmichael1/burro/23