Add support for relative units
This allows users to change parameters relative to what they currently
for. As an example, if the current point size is 12, then .pt_size[+8]
changes the point size to 20.

Relative units can only be used where it makes sense to use them :)
Trying to use them in, e.g., the .font command will cause an error. They
will be ignored if used in .vspace or .hspace (in a sense, the spacing
commands are already relative to the cursor). However, relative units
should work in pretty much any numeric context, such as .pt_size,
.margins, or .leading.

Implements: https://todo.sr.ht/~reesmichael1/burro/26