First pass at doing Mom's tabs demo in Burro
There's a lot of work to be done here (with plenty of TODOs and todo!()
left in the codebase as markers), but this is a good checkpoint commit.
Tabs are starting to become useful!

One limitation that might not be marked explicitly: There's still no
distinction between quad and non-quad mode.