Add support for tab stops
This was a significant amount of work done in the tabs branch. I'll
leave that branch alive for posterity so that all of those commits and
history arent lost.

The only thing missing from this commit is support for using percentages
as the unit for tab indents/lengths. This will be very nice to have, but
will also be a pretty substantial change to the internals, so I decided
to wait until tabs were in master to start on that work.