Properly consider letter spacing in line width
This fixes a long standing bug where changing the letter spacing could
push text into the margin.

However, this revealed a major problem in the layout algorithm: emitting
one word at a time works fine normally, but when we try to change
parameters in the middle of the line while right aligned, the text is
emitted too far to the right, since the cursor is being advanced from
the right margin.

Fixes: https://todo.sr.ht/~reesmichael1/burro/21