Use emit chunks instead of words in layout
While the summary might sound innocuous, this is a substantial rewrite
of the layout algorithm, which fixes the major layout bug. Before, we'd
emit one word at a time and then finalize their positions once
"something" changed. However, this led to us putting words on the page
too soon in some cases.

Most notably, when right aligned, parameter changes caused the first
part of the line to be emitted along the right margin, and then the rest
of the line to be emitted on top of that. By using emit chunks, we now
only finalize the word position once the entire line is done and ready
to be emitted.

Fixes: https://todo.sr.ht/~reesmichael1/burro/34
Fixes: https://todo.sr.ht/~reesmichael1/burro/35